fix(cta): create custom hook for video data
